refactor#1_styling_component_back
