+++ db.json
@@ -8,17 +8,17 @@
     {
       "id": "1706579849272",
       "todoText": "gdgdg",
-      "isCompleted": false
-    },
-    {
-      "id": "1706579851216",
-      "todoText": "nvnvnv",
-      "isCompleted": false
+      "isCompleted": true
     },
     {
       "id": "1706582030890",
       "todoText": "asfs",
       "isCompleted": true
+    },
+    {
+      "id": "1706583541564",
+      "todoText": "ad",
+      "isCompleted": false
     }
   ]
 }
