+++ src/API/js/addItem.js
@@ -1,4 +1,4 @@
-import { VITE_APP_KEY } from "./apiKey";
+import { VITE_APP_KEY } from "../key/apiKey";
 export async function addItem(
   plusTodoText,
   todoArr,
