+++ src/API/js/completeItem.js
@@ -1,4 +1,4 @@
-import { VITE_APP_KEY } from "./apiKey";
+import { VITE_APP_KEY } from "../key/apiKey";
 export async function completeItem(todoItemId, isCompleted) {
   await fetch(`${VITE_APP_KEY}/todo/${todoItemId}`, {
     method: "PATCH",
