+++ src/API/js/deleteItem.js
@@ -1,4 +1,4 @@
-import { VITE_APP_KEY } from "./apiKey";
+import { VITE_APP_KEY } from "../key/apiKey";
 export async function deleteItem(todoItemId, todoArr, setTodoArr) {
   const isSuccess = await fetch(`${VITE_APP_KEY}/todo/${todoItemId}`, {
     method: "DELETE",
