+++ src/component/AddForm/AddForm.jsx
@@ -2,13 +2,6 @@ import styled from "styled-components";
 import AddButton from "../button/AddButton/AddButton";
 import TodoInput from "../Input/TodoInput/TodoInput";
 
-const FormStyle = styled.div`
-  width: 397px;
-  height: 44px;
-  display: flex;
-  justify-content: space-between;
-`;
-
 const AddForm = ({ addBtnFun }) => {
   return (
     <>
@@ -21,3 +14,10 @@ const AddForm = ({ addBtnFun }) => {
 };
 
 export default AddForm;
+
+const FormStyle = styled.div`
+  width: 397px;
+  height: 44px;
+  display: flex;
+  justify-content: space-between;
+`;
