+++ src/component/Input/TodoInput/TodoInput.jsx
@@ -1,18 +1,6 @@
 import { useContext } from "react";
 import styled from "styled-components";
-
 import { PlusTodoText, SetPlusTodoText } from "../../../pages/MainPage";
-
-const TodoInputStyle = styled.input`
-  width: 324px;
-  height: 44px;
-  font-family: "Pretendard-Medium";
-  font-size: 20px;
-  padding-left: 20px;
-  box-sizing: border-box;
-  border: 1px solid #7e6363;
-  border-radius: 8px;
-`;
 
 const TodoInput = () => {
   const plusTodoText = useContext(PlusTodoText);
@@ -34,3 +22,14 @@ const TodoInput = () => {
 };
 
 export default TodoInput;
+
+const TodoInputStyle = styled.input`
+  width: 324px;
+  height: 44px;
+  font-family: "Pretendard-Medium";
+  font-size: 20px;
+  padding-left: 20px;
+  box-sizing: border-box;
+  border: 1px solid #7e6363;
+  border-radius: 8px;
+`;
