+++ src/component/button/AddButton/AddButton.jsx
@@ -1,21 +1,5 @@
 import styled from "styled-components";
 
-const AddButtonStyle = styled.button`
-  width: 65px;
-  height: 44px;
-  border-radius: 8px;
-  background-color: #a87c7c;
-  color: white;
-  font-size: 20px;
-  font-family: "Pretendard-bold";
-  text-align: center;
-  line-height: 44px;
-  border: none;
-  &:hover {
-    background-color: #ebd9b4;
-  }
-`;
-
 const AddButton = ({ addBtnFun }) => {
   return (
     <>
@@ -31,3 +15,19 @@ const AddButton = ({ addBtnFun }) => {
 };
 
 export default AddButton;
+
+const AddButtonStyle = styled.button`
+  width: 65px;
+  height: 44px;
+  border-radius: 8px;
+  background-color: #a87c7c;
+  color: white;
+  font-size: 20px;
+  font-family: "Pretendard-bold";
+  text-align: center;
+  line-height: 44px;
+  border: none;
+  &:hover {
+    background-color: #ebd9b4;
+  }
+`;
