+++ src/component/header/Header.jsx
@@ -1,14 +1,5 @@
 import styled from "styled-components";
 
-const Title = styled.div`
-  width: 100%;
-  height: auto;
-  font-family: "Pretendard-bold";
-  color: #503c3c;
-  font-size: 30px;
-  text-align: center;
-`;
-
 const Header = ({ children }) => {
   return (
     <>
@@ -18,3 +9,12 @@ const Header = ({ children }) => {
 };
 
 export default Header;
+
+const Title = styled.div`
+  width: 100%;
+  height: auto;
+  font-family: "Pretendard-bold";
+  color: #503c3c;
+  font-size: 30px;
+  text-align: center;
+`;
